Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -2,7 +2,7 @@
 int main(){
     float a,b;
     scanf("%.2f %.2f",&a,&b);
-    float Product;
+    double Product;
     Product=a*b;
     printf("Product: %.2f",Product);
     return 0;
